modules/locations2kafka/requirements.txt
grpcio==1.41.1
grpcio-reflection==1.41.1
grpcio-tools==1.41.1
kafka-python==2.0.2
protobuf==3.19.1
six==1.16.0
